+++ src/test/groovy/ru/vyarus/guice/persist/orient/study/index/fieldsorder/FieldsOrderRepository.groovy
@@ -4,6 +4,7 @@ import com.google.inject.ProvidedBy
 import com.google.inject.internal.DynamicSingletonProvider
 import com.google.inject.persist.Transactional
 import com.orientechnologies.orient.core.record.impl.ODocument
+import ru.vyarus.guice.persist.orient.model.VersionedEntity
 import ru.vyarus.guice.persist.orient.repository.command.query.Query
 import ru.vyarus.guice.persist.orient.support.repository.mixin.crud.ObjectCrud
 
@@ -13,7 +14,7 @@ import ru.vyarus.guice.persist.orient.support.repository.mixin.crud.ObjectCrud
  */
 @Transactional
 @ProvidedBy(DynamicSingletonProvider)
-interface FieldsOrderRepository extends ObjectCrud<FOTest> {
+interface FieldsOrderRepository extends ObjectCrud<VersionedEntity> {
 
     @Query("explain select from FOTest where foo=? and bar=?")
     ODocument sameOrder()
@@ -23,4 +24,7 @@ interface FieldsOrderRepository extends ObjectCrud<FOTest> {
 
     @Query("explain select from FOTest where foo=?")
     ODocument foo()
+
+    @Query("explain select from FOTest2 where foo=?")
+    ODocument foo2()
 }